+++ src/BranchHistory.cpp
@@ -35,7 +35,7 @@ BranchHistory::~BranchHistory(void)
 void BranchHistory::printBranchHistory(void)
 {
 
-    int nEvents = eventsOnBranch.size();
+    int nEvents = (int)eventsOnBranch.size();
     std::cout << "nodeEvent: " << nodeEvent << "\tancestorEvent: " <<
         ancestralNodeEvent << std::endl;
